--- project/src/app/app.component.css
@@ -1,4 +1,0 @@
-#colorstar { color: #ee8b2d; }
-.badForm { color: #FF0000; }
-.goodForm { color: #00FF00; }
-.evaluation { margin-left:30px; }
