+++ project/src/app/app.component.html
@@ -7,11 +7,11 @@
         <span class="icon-bar"></span>
         <span class="icon-bar"></span>
       </button>
-      <a class="navbar-brand" href="#">Online Course Tracker</a>
+      <a class="navbar-brand" href=".">Online Course Tracker</a>
     </div>    
     <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
       <ul class="nav navbar-nav">
-        <li class="active"><a href="#">Home <span class="sr-only">(current)</span></a></li>
+        <li class="active"><a href=".">Home <span class="sr-only">(current)</span></a></li>
         <li><a href="#">About</a></li>       
       </ul>
       <ul class="nav navbar-nav navbar-right">
@@ -156,14 +156,12 @@
 </div>
 <!-- EDIT FORM ENDS -->
 
-
-
 <!-- EXTEND SECTION Starts -->
 <div *ngIf="appState == 'extend'" class="container-fluid">
   <div class="row">
-    <div class="col-md-12 jumbotron">
+    <div class="col-md-12">
       <p>
-        <button (click)="changeState('default')" class="btn btn-warning">Go Back</button>
+        <button (click)="changeState('default')" class="btn btn-warning" style="float:right; margin-bottom: 10px;"> X </button>
       </p>
     </div>
   </div>
@@ -252,62 +250,3 @@
     </ul>
   </div>
 -->
-
-<style>
-  #colorstar { color: #ee8b2d; }
-  .badForm { color: #FF0000; }
-  .goodForm { color: #00FF00; }
-  .evaluation { margin-left:30px; } 
-</style>
-
-                <section>
-  <div class="container">
-    <h2>Course Evaluation<small> Description </small></h2>
-  </div>
-  <div class="container">
-    <h3>[Victor] Udemy Course</h3>
-    <div class="row lead evaluation">
-      <div id="colorstar" class="starrr ratable"></div>
-      <span id="count">0</span> star(s) - <span id="meaning"> </span>
-      <div class='indicators' style="display:none">
-        <div id='textwr'>What happend?</div>
-        <input id="rate[]" name="rate[]" type="text" placeholder="" class="form-control input-md" style="display:none;">
-        <input id="rating[]" name="rating[]" type="text" placeholder="" class="form-control input-md rateval" style="display:none;">
-        <span class="button-checkbox">
-          <button type="button" class="btn criteria" data-color="info">Engagement</button>
-            <input type="checkbox" class="hidden"  />
-          </span>
-        <span class="button-checkbox">
-          <button type="button" class="btn criteria" data-color="info">Value</button>
-            <input type="checkbox" class="hidden"  />
-          </span>
-        <span class="button-checkbox">
-          <button type="button" class="btn criteria" data-color="info">Knowledge</button>
-            <input type="checkbox" class="hidden"  />
-          </span>
-      </div>
-    </div>
-    <h3>[Daniel] Code School</h3>
-    <div class="row lead evaluation">
-      <div id="colorstar" class="starrr ratable"></div>
-      <span id="count">0</span> star(s) - <span id="meaning"> </span>
-      <div class='indicators' style="display:none">
-        <div id='textwr'>What happened?</div>
-        <input id="rate[]" name="rate[]" type="text" placeholder="" class="form-control input-md" style="display:none;">
-        <input id="rating[]" name="rating[]" type="text" placeholder="" class="form-control input-md rateval" style="display:none;">
-        <span class="button-checkbox">
-          <button type="button" class="btn" data-color="info">Engagement</button>
-            <input type="checkbox" class="hidden" />
-          </span>
-        <span class="button-checkbox">
-          <button type="button" class="btn" data-color="info">Value</button>
-            <input type="checkbox" class="hidden" />
-          </span>
-        <span class="button-checkbox">
-          <button type="button" class="btn" data-color="info">Knowledge</button>
-            <input type="checkbox" class="hidden" />
-          </span>
-      </div>
-    </div>
-  </div>
-</section>
