+++ project/src/app/components/review.component.css
@@ -1,0 +1,4 @@
+#colorstar { color: #ee8b2d; }
+.badForm { color: #FF0000; }
+.goodForm { color: #00FF00; }
+.evaluation { margin-left:30px; }
