+++ project/src/app/components/review.component.html
@@ -1,0 +1,51 @@
+<section>
+  <div class="container">
+    <h2>Course Evaluation<small> {{ course.school }} </small></h2>
+  </div>
+  <div class="container">
+    <h3>[Victor] Udemy Course</h3>
+    <div class="row lead evaluation">
+      <div id="colorstar" class="starrr ratable"></div>
+      <span id="count">0</span> star(s) - <span id="meaning"> </span>
+      <div class='indicators' style="display:none">
+        <div id='textwr'>What happend?</div>
+        <input id="rate[]" name="rate[]" type="text" placeholder="" class="form-control input-md" style="display:none;">
+        <input id="rating[]" name="rating[]" type="text" placeholder="" class="form-control input-md rateval" style="display:none;">
+        <span class="button-checkbox">
+          <button type="button" class="btn criteria" data-color="info">Engagement</button>
+            <input type="checkbox" class="hidden"  />
+          </span>
+        <span class="button-checkbox">
+          <button type="button" class="btn criteria" data-color="info">Value</button>
+            <input type="checkbox" class="hidden"  />
+          </span>
+        <span class="button-checkbox">
+          <button type="button" class="btn criteria" data-color="info">Knowledge</button>
+            <input type="checkbox" class="hidden"  />
+          </span>
+      </div>
+    </div>
+    <h3>[Daniel] Code School</h3>
+    <div class="row lead evaluation">
+      <div id="colorstar" class="starrr ratable"></div>
+      <span id="count">0</span> star(s) - <span id="meaning"> </span>
+      <div class='indicators' style="display:none">
+        <div id='textwr'>What happened?</div>
+        <input id="rate[]" name="rate[]" type="text" placeholder="" class="form-control input-md" style="display:none;">
+        <input id="rating[]" name="rating[]" type="text" placeholder="" class="form-control input-md rateval" style="display:none;">
+        <span class="button-checkbox">
+          <button type="button" class="btn" data-color="info">Engagement</button>
+            <input type="checkbox" class="hidden" />
+          </span>
+        <span class="button-checkbox">
+          <button type="button" class="btn" data-color="info">Value</button>
+            <input type="checkbox" class="hidden" />
+          </span>
+        <span class="button-checkbox">
+          <button type="button" class="btn" data-color="info">Knowledge</button>
+            <input type="checkbox" class="hidden" />
+          </span>
+      </div>
+    </div>
+  </div>
+</section>
