+++ project/src/app/components/review.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import 'rxjs/add/operator/map';
+
+@Component({
+    selector: 'review',
+    templateUrl: './app/components/review.component.html',
+    styleUrls: ['./app/components/review.component.css']
+})
+
+export class ReviewComponent {
+  
+}
